+++ heylogs-api/src/main/java/internal/heylogs/github/GitHub.java
@@ -3,8 +3,13 @@ package internal.heylogs.github;
 import lombok.NonNull;
 import nbbrd.design.DirectImpl;
 import nbbrd.heylogs.spi.Forge;
+import nbbrd.io.http.URLQueryBuilder;
 import nbbrd.io.text.Parser;
 import nbbrd.service.ServiceProvider;
+
+import java.io.UncheckedIOException;
+import java.net.MalformedURLException;
+import java.net.URL;
 
 @DirectImpl
 @ServiceProvider
@@ -23,5 +28,19 @@ public final class GitHub implements Forge {
     @Override
     public boolean isCompareLink(@NonNull CharSequence text) {
         return Parser.of(GitHubCompareLink::parse).parseValue(text).isPresent();
+    }
+
+    @Override
+    public @NonNull URL getBaseURL(@NonNull CharSequence text) {
+        GitHubCompareLink compareLink = GitHubCompareLink.parse(text);
+        try {
+            return URLQueryBuilder
+                    .of(compareLink.getBase())
+                    .path(compareLink.getOwner())
+                    .path(compareLink.getRepo())
+                    .build();
+        } catch (MalformedURLException ex) {
+            throw new UncheckedIOException(ex);
+        }
     }
 }