+++ heylogs-api/src/main/java/nbbrd/heylogs/spi/Forge.java
@@ -5,15 +5,22 @@ import nbbrd.service.Quantifier;
 import nbbrd.service.ServiceDefinition;
 import nbbrd.service.ServiceId;
 
+import java.net.URL;
+
 @ServiceDefinition(
         quantifier = Quantifier.MULTIPLE
 )
 public interface Forge {
 
     @ServiceId(pattern = ServiceId.KEBAB_CASE)
-    @NonNull String getForgeId();
+    @NonNull
+    String getForgeId();
 
-    @NonNull String getForgeName();
+    @NonNull
+    String getForgeName();
 
     boolean isCompareLink(@NonNull CharSequence text);
+
+    @NonNull
+    URL getBaseURL(@NonNull CharSequence text);
 }